split core process into separate class
