+++ vllm/v1/engine/llm_engine.py
@@ -17,8 +17,8 @@ from vllm.sampling_params import SamplingParams
 from vllm.usage.usage_lib import UsageContext
 from vllm.v1.engine.detokenizer import Detokenizer
 from vllm.v1.engine.llm_engine_core import LLMEngineCore
-from vllm.v1.engine.protocol import LLMEngineProtocol
 from vllm.v1.engine.processor import Processor
+from vllm.v1.engine.protocol import LLMEngineProtocol
 from vllm.v1.executor.gpu_executor import GPUExecutor
 
 logger = init_logger(__name__)
